+++ src/index.js
@@ -1,4 +1,5 @@
+import Animate from './Animate'
 import Collapse from './Collapse'
 import Fluid from './Fluid'
 
-export { Collapse, Fluid }
+export { Animate, Collapse, Fluid }