+++ CHANGELOG.md
@@ -1,4 +1,12 @@
 ## CHANGELOG
 
+### 0.2.0
+Added `Animate` component
+
+Added `render` prop to `Collapse` to allow custom child component
+
+### 0.1.1
+Fix `package.json` for NPM
+
 ### 0.1.0
 Initial release